Timezone can be + 30 mins
Fixes #13887

+++ includes/wc-formatting-functions.php
@@ -612,7 +612,7 @@ function wc_timezone_offset() {
 		$timezone_object = new DateTimeZone( $timezone );
 		return $timezone_object->getOffset( new DateTime( 'now' ) );
 	} else {
-		return intval( get_option( 'gmt_offset', 0 ) ) * HOUR_IN_SECONDS;
+		return floatval( get_option( 'gmt_offset', 0 ) ) * HOUR_IN_SECONDS;
 	}
 }
